Backend/requirements.txt
fastapi==0.115.5
uvicorn==0.32.0
python-multipart==0.0.9
Pillow==10.4.0
numpy==2.1.2
sqlmodel==0.0.22
pydantic==2.9.2
python-dotenv==1.0.1